Seeders agregados (Users, News, Services)

+++ app/Http/Controllers/ServicesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Service;
+use App\Models\Suscription;
 
 class ServicesController extends Controller
 {
@@ -38,14 +39,23 @@ class ServicesController extends Controller
         $service = Service::find($id);
 
         if ($service) {
+            $hasSubscription = Suscription::where('service_id', $service->id)
+                ->exists();
+
+            if ($hasSubscription) {
+                return redirect()->route('services')->with('error', 'No puedes eliminar este servicio porque existe una suscripción activa.');
+            }
+
             $service->delete();
-            return redirect()->route('services')->with('success', 'Servicio eliminado');
+            return redirect()->route('services')->with('success', 'Servicio eliminado.');
         }
-        // por si no existe el service
+
         return redirect()->route('services.index')->with('error', 'Servicio no encontrado.');
     }
 
-    public function editService(Request $request, $id){
+
+    public function editService(Request $request, $id)
+    {
         //valida la entrada y manda mensajes personalizados de error
         $request->validate([
             'name' => 'required|string|max:255',
